python-webhook/requirements.txt
Flask==2.2.2
kubernetes==26.1.0
PyYAML==6.0
gunicorn==20.1.0